// src/app/models/agent-sign-up.ts
export interface AgentSignUp {
    agentId?: string;
    confirmed?: string;
    flag?: string;
    lotId?: string;
    cin: string;
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    city: string;
    mobile: string;
    country: string;
    street: string;
    nameAgency: string;
}